# openapi.json
{
  "components": {
    "schemas": {
...
      "AutomationClientCreateRequest": {
        "additionalProperties": false,
        "properties": {
          "client_id": {
            "pattern": "^[0-9a-f]{8}-[0-9a-f]{4}-4[0-9a-f]{3}-[89ab][0-9a-f]{3}-[0-9a-f]{12}$",
            "title": "Client Id",
            "type": "string"
          },
          "name": {
            "maxLength": 160,
            "minLength": 1,
            "title": "Name",
            "type": "string"
          },
          "redirect_uris": {
            "items": {
              "type": "string"
            },
            "maxItems": 10,
            "minItems": 1,
            "title": "Redirect Uris",
            "type": "array"
          },
          "scopes": {
            "items": {
              "enum": [
                "app.read",
                "app.write",
                "app.run",
                "app.cancel",
                "app.credentials.read",
                "app.credentials.write",
                "manager.read",
                "manager.runtime",
                "manager.mutate",
                "app.admin"
              ],
              "type": "string"
            },
            "minItems": 1,
            "title": "Scopes",
            "type": "array"
          }
        },
        "required": [
          "client_id",
          "name",
          "redirect_uris",
          "scopes"
        ],
        "title": "AutomationClientCreateRequest",
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "bearerToken": {
        "scheme": "bearer",
        "type": "http"
      },
      "cookieAuth": {
        "in": "cookie",
        "name": "session",
        "type": "apiKey"
      },
      "nativeOAuth": {
        "flows": {
          "authorizationCode": {
            "authorizationUrl": "/api/v1/auth/automation/authorize",
            "scopes": {
              "app.cancel": "Cancel durable application work.",
              "app.credentials.read": "Inspect credential status.",
              "app.credentials.write": "Change credential references.",
              "app.read": "Read application state.",
              "app.run": "Start durable application work.",
              "app.write": "Change reversible application state.",
              "manager.mutate": "Execute approved Manager plans.",
              "manager.read": "Inspect Manager state and plans.",
              "manager.runtime": "Control managed runtime services."
            },
            "tokenUrl": "/api/v1/auth/automation/token"
          }
        },
        "type": "oauth2"
      }
    }
  },
  "info": {
    "title": "Pandrator API",
    "version": "1.0.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/auth/automation-clients": {
      "get": {
        "operationId": "listAutomationClients",
        "responses": {
          "200": {
            "description": "Automation clients"
          }
        }
      },
      "post": {
        "operationId": "registerAutomationClient",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AutomationClientCreateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Registered"
          }
        }
      }
    },
    "/api/v1/auth/automation-clients/{clientId}": {
      "delete": {
        "operationId": "revokeAutomationClient",
        "parameters": [
          {
            "in": "path",
            "name": "clientId",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Revoked"
          }
        }
      }
    },
    "/api/v1/auth/automation/authorize": {
      "get": {
        "description": "Trusted owner consent with exact redirect matching and mandatory S256 PKCE.",
        "operationId": "authorizeAutomationClient",
        "responses": {
          "200": {
            "description": "Consent page"
          },
          "302": {
            "description": "Authorization response"
          }
        }
      }
    },
...
  },
  "servers": [
    {
      "url": "/"
    }
  ]
}